Implemented Selection Grid for Adding a Script and Rearrange Files
Added code to the RPGTool script which allows checks which button is selected from the selection grid on tab 2 and adds a script the the newly create object when the “Create Character” button is pushed.
Rearranged files again due to scripts being prevented from getting added to an object if in the “Editor” folder.