Implemented Basic Movement With Mouse
Added new command which gets the position of the mouse when used in the game window and moves a game object which the character will have velocity applied to it towards the new game object.
Locked the rigidbody from moving on the y axis as a temporary fix until pathfinding is implemented.
Added TraversableGround tag which is applied to any ground that the characters should be able to traverse.