Added Character Tab Fields
Added buttons for creating both player and hostile character which creates a new character using the fields and stores them in the CharacterList (Characters are not stored currently).
Changed m_levelCurve to m_experienceCurve as this depicts what it represents better.
Create method to reset all fields to the default values they should be.